Changes to turn into a package
Modified files to get them ready to be a package.  Added a SimpleTabItem and a number of changes to allow you to pass basic info to the widget and have it be dynamic, allowing any number of items to be put into the navigation bar.

One issue is the position of the FAB on load to the correct spot.